[REFACTORING] 도메인에 맞게 코드 위치 변경
