--- naegamaja/src/main/java/naegamaja_server/naegamaja/domain/chat/service/chatLogService.java
@@ -1,4 +1,0 @@
-package naegamaja_server.naegamaja.domain.chat.service;
-
-public class chatLogService {
-}
